output/json: Partial fix for json data race.
The previous implementation makes a copy of the Metric struct
where it generates a data race on the Observed field that is written
from the metrics.Engine.

The current implementation reduces the fields copied resolving
the simplest race on Observed. Observed field is not useful for the JSON output.

Thresholds shouldn't be racy because the output mostly read
"immutable" values.

The Tainted is still racy and they should be refactored
when the refactor on time series and metrics will be completed.

The performance seems not impacted, it is aligned with the
result from a5146cf7099cfc5b4ec8938c667c507ce373aff9.

BenchmarkFlushMetrics-8   	      61	  16767247 ns/op	 2526610 B/op	   20067 allocs/op